This is currently useless
In the context of 67-475, we don't need to waste server-side memory 'saving' this data for no one to use. In the context of an actual company, this data would need to instead be persisted in some form of permanent data storage, be it Cassandra, S3, MongoDB, or any of the literally hundreds of solutions out there.